# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.User:
    properties:
      avatar:
        description: 头像
        type: string
      birthdate:
        description: 生日
        type: string
      created_at:
        type: string
      email:
        description: 邮箱
        type: string
      gender:
        description: 性别
        type: string
      id:
        example: "0"
        type: string
      lastLoginAt:
        description: 最近登录时间
        type: string
      nickname:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号
        type: string
      role:
        description: 角色
        type: string
      status:
        description: 状态
        type: string
      updated_at:
        type: string
      username:
        description: 用户名
        type: string
    type: object
info:
  contact:
    email: iszhounie@gmail.com
    name: Zoey
    url: http://localhost:8080
  description: IKunCoffee 接口文档
  title: IKunCoffee API DOCS
  version: "1.0"
paths:
  /api/v1/users:
    delete:
      description: 根据ID删除用户
      parameters:
      - description: ids
        in: query
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: 删除用户
      tags:
      - users
    get:
      description: 创建用户
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: 创建用户
      tags:
      - users
  /api/v1/users/:id:
    post:
      description: 根据ID获取用户
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: 获取用户
      tags:
      - users
swagger: "2.0"